Flatten LSPViewEventListener inheritance hierarchy
